Kinda a MVP Stage
It works and such is where I am for now, I want to add controller support, and simple audio but for now this should be good. You immediately die when touched by spikes or a bat and the camera will fade out to make this clear. Collecting all coins will make the screen go white for now as placeholder